// src/features/Rooms/Types/RoomType.ts
export type Reservation = {
  observations: string;
  endTime: string;
  startTime: string;
  date: string;
  name: string;
  EventId?: number;
  courseId?: number;
  userCedula: string;
  roomId: number;
  reason: string;
  personNumber: string;
};
export type Reserve = {
  observations: string;
  endTime: string;
  startTime: string;
  reservationDate: string;
  date: string;
  name: string;
  EventId?: number;
  courseId?: number;
  userCedula: string;
  roomId: number;
  reason: string;
  personNumber: string;
};